Player animation pre-overhaul push
Discovered that the dynamic animation system I implemented for both the player and the enemy can only be implemented in the Unity editor. Tried to publish a build, but failed.

I am now doing a complete overhaul in which I will simply prepare the animation assets beforehand and use a blend tree in the animator. I am also considering using an animator override controller, but I feel like it's somewhat of an overkill for such a small part of the game.