Inventory implementation
Starting the inventory system. The system will be a basic system, wherein the player can pick-up an item and use it. No need to mix and match, but is building the foundation for a possible upgrade to this kind of a system in the long run.

Just implemented the picking up. The UI needs to be fixed next.